Enable mapping IDs to usernames.
In order to display a list of current players in a game window,
the server needs to keep track of what IDs map to what usernames.
The appropriate changes have been made to enable this.

As a consequence, the dummy proxy should now be passed around
through the different parts of the program to keep the same
instance around, though this will not be necessary with a full
client-server implementation. This is simply emulating the effect
of having a single server.