Link universe and sector views.
When a new sector is selected in the universe view, the sector
view needs to update based on the selection. When the universe
view detects a sector selection, it will tell the sector view of
the new sector and the sector view will repaint itself after
setting the new sector internally.