Do drawing for ships.
A couple of changes had to be made to the mapping of players to
ships and ships to players. Players are now represented by their
username and not a Player object, which was a fancy way of
representing these values.

This also makes it possible to discover what user is in what
ship and discover alliances to properly display alliances with
colors. The player's ship is yellow; enemies are red; friendly
bases are green.