Fix adding and removing players.
When adding a player, we first need to check to see if the player
has already played. If so, there is no need to insert the key into
the map again. If anything, it's a waste of time since they're
already in the map.

Players should be removed by the Game member function. This is now
fixed by actually going through the ArrayList (slow) and finding
the player to remove. Trying to match the Player object with a
string will not work.