Fix game over sequence.
After the player's ship is destroyed, it is removed from the list
of ships currently in the game. The list is used to draw objects
appropriately due to the need of differentiating alliances. The
drawing should not occur if the game over condition is met.

Also, this change removes the extra 'Game Over' dialog box that
would appear upon dying.